% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figure1a_rhat_ecdf}
\alias{create_figure1a_rhat_ecdf}
\title{Create Figure 1a: R-hat ECDF by sample size}
\usage{
create_figure1a_rhat_ecdf(plot_data)
}
\arguments{
\item{plot_data}{Prepared plot data from prepare_plot_data()}
}
\description{
Create Figure 1a: R-hat ECDF by sample size
}
