% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figure1b_ess_ridges}
\alias{create_figure1b_ess_ridges}
\title{Create Figure 1b: ESS ridges by scenario}
\usage{
create_figure1b_ess_ridges(plot_data)
}
\arguments{
\item{plot_data}{Prepared plot data from prepare_plot_data()}
}
\description{
Create Figure 1b: ESS ridges by scenario
}
