% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figure2_coverage}
\alias{create_figure2_coverage}
\title{Create Figure 2: Coverage of 95\% CIs}
\usage{
create_figure2_coverage(plot_data)
}
\arguments{
\item{plot_data}{Prepared plot data from prepare_plot_data()}
}
\description{
Create Figure 2: Coverage of 95\% CIs
}
