% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figure3_10_zimphia_ess_per_sec}
\alias{create_figure3_10_zimphia_ess_per_sec}
\title{ZIMPHIA ESS per second by sampler}
\usage{
create_figure3_10_zimphia_ess_per_sec(zimphia_dir = "mcmc_outputs/zimphia")
}
\arguments{
\item{zimphia_dir}{Directory containing ZIMPHIA outputs.}
}
\value{
ggplot object
}
\description{
ZIMPHIA ESS per second by sampler
}
