% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figure3_11_zimphia_runtime_bars}
\alias{create_figure3_11_zimphia_runtime_bars}
\title{ZIMPHIA runtime comparison}
\usage{
create_figure3_11_zimphia_runtime_bars(zimphia_dir = "mcmc_outputs/zimphia")
}
\arguments{
\item{zimphia_dir}{Directory containing ZIMPHIA outputs.}
}
\value{
ggplot object
}
\description{
ZIMPHIA runtime comparison
}
