% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figure3_7_zimphia_convergence_traces}
\alias{create_figure3_7_zimphia_convergence_traces}
\title{ZIMPHIA convergence diagnostics via trace plots}
\usage{
create_figure3_7_zimphia_convergence_traces(
  zimphia_dir = "mcmc_outputs/zimphia"
)
}
\arguments{
\item{zimphia_dir}{Directory containing ZIMPHIA outputs.}
}
\value{
ggplot object
}
\description{
ZIMPHIA convergence diagnostics via trace plots
}
