% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figure3_8_zimphia_posterior_forest}
\alias{create_figure3_8_zimphia_posterior_forest}
\title{Figure 3.8 — Posterior distributions for seroconversion model}
\usage{
create_figure3_8_zimphia_posterior_forest(zimphia_dir = "mcmc_outputs/zimphia")
}
\arguments{
\item{zimphia_dir}{Directory containing ZIMPHIA outputs.}
}
\value{
ggplot object
}
\description{
Figure 3.8 — Posterior distributions for seroconversion model
}
