% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figure3_9_zimphia_beta_comparison}
\alias{create_figure3_9_zimphia_beta_comparison}
\title{ZIMPHIA beta posterior overlap}
\usage{
create_figure3_9_zimphia_beta_comparison(zimphia_dir = "mcmc_outputs/zimphia")
}
\arguments{
\item{zimphia_dir}{Directory containing ZIMPHIA outputs.}
}
\value{
ggplot object
}
\description{
ZIMPHIA beta posterior overlap
}
