% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figure3a_bias}
\alias{create_figure3a_bias}
\title{Create Figure 3a: Bias by parameter}
\usage{
create_figure3a_bias(plot_data)
}
\arguments{
\item{plot_data}{Prepared plot data from prepare_plot_data()}
}
\description{
Create Figure 3a: Bias by parameter
}
