% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figure4_ess_per_sec}
\alias{create_figure4_ess_per_sec}
\title{Create Figure 4: ESS per second}
\usage{
create_figure4_ess_per_sec(plot_data)
}
\arguments{
\item{plot_data}{Prepared plot data from prepare_plot_data()}
}
\description{
Create Figure 4: ESS per second
}
