% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figure7_means_agreement}
\alias{create_figure7_means_agreement}
\title{Create Figure 7: Agreement of posterior means}
\usage{
create_figure7_means_agreement(plot_data)
}
\arguments{
\item{plot_data}{Prepared plot data from prepare_plot_data()}
}
\description{
Create Figure 7: Agreement of posterior means
}
