% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figureA2_ci_matrix}
\alias{create_figureA2_ci_matrix}
\title{Create Figure A2: CI matrix}
\usage{
create_figureA2_ci_matrix(plot_data)
}
\arguments{
\item{plot_data}{Prepared plot data from prepare_plot_data()}
}
\description{
Create Figure A2: CI matrix
}
