% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figureB3_ci_width_violin}
\alias{create_figureB3_ci_width_violin}
\title{Create Figure B3: CI width violin plots}
\usage{
create_figureB3_ci_width_violin(plot_data)
}
\arguments{
\item{plot_data}{Prepared plot data from prepare_plot_data()}
}
\description{
Create Figure B3: CI width violin plots
}
