% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_default_params}
\alias{get_default_params}
\title{Simulation utilities and orchestration}
\usage{
get_default_params()
}
\value{
A list of default parameter values
}
\description{
Simulation utilities and orchestration
}
