% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{get_plot_res}
\alias{get_plot_res}
\title{Extract the primary plotting tibble from prepared data}
\usage{
get_plot_res(plot_data)
}
\description{
Extract the primary plotting tibble from prepared data
}
\keyword{internal}
