% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_zimphia_paths}
\alias{get_zimphia_paths}
\title{Get canonical ZIMPHIA output paths}
\usage{
get_zimphia_paths(base_dir = "mcmc_outputs/zimphia")
}
\description{
Get canonical ZIMPHIA output paths
}
\keyword{internal}
