% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalesce operator (internal)}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Second value (returned if x is NULL)}
}
\value{
x if not NULL, otherwise y
}
\description{
Null coalesce operator (internal)
}
\keyword{internal}
