% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{load_zimphia_diagnostics}
\alias{load_zimphia_diagnostics}
\title{Load ZIMPHIA diagnostic tables}
\usage{
load_zimphia_diagnostics(base_dir = "mcmc_outputs/zimphia")
}
\description{
Load ZIMPHIA diagnostic tables
}
\keyword{internal}
