% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{load_zimphia_draws}
\alias{load_zimphia_draws}
\title{Internal helper: load ZIMPHIA draws}
\usage{
load_zimphia_draws(method = c("hmc", "mh"), base_dir = "mcmc_outputs/zimphia")
}
\description{
Internal helper: load ZIMPHIA draws
}
\keyword{internal}
