% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{load_zimphia_mh_fit}
\alias{load_zimphia_mh_fit}
\title{Load MH fit object for acceptance statistics}
\usage{
load_zimphia_mh_fit(base_dir = "mcmc_outputs/zimphia")
}
\description{
Load MH fit object for acceptance statistics
}
\keyword{internal}
