% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{load_zimphia_prepared_data}
\alias{load_zimphia_prepared_data}
\title{Load analytic ZIMPHIA dataset}
\usage{
load_zimphia_prepared_data(base_dir = "mcmc_outputs/zimphia")
}
\description{
Load analytic ZIMPHIA dataset
}
\keyword{internal}
