% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{load_zimphia_summaries}
\alias{load_zimphia_summaries}
\title{Load ZIMPHIA summary tables}
\usage{
load_zimphia_summaries(base_dir = "mcmc_outputs/zimphia")
}
\description{
Load ZIMPHIA summary tables
}
\keyword{internal}
