% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{summarise_zimphia_missing_flow}
\alias{summarise_zimphia_missing_flow}
\title{Summarise ZIMPHIA missingness flow across inclusion criteria}
\usage{
summarise_zimphia_missing_flow(
  adultbio_file = file.path("ZIMPHIA", "ZIMPHIA 2020 Datasets (CSV)",
    "zimphia2020adultbio.csv"),
  adultind_file = file.path("ZIMPHIA", "ZIMPHIA 2020 Datasets (CSV)",
    "zimphia2020adultind.csv")
)
}
\description{
Summarise ZIMPHIA missingness flow across inclusion criteria
}
\keyword{internal}
