% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_3_8_parameter_estimates}
\alias{table_3_8_parameter_estimates}
\title{Table 3.8 -- Parameter estimates from ZIMPHIA application}
\usage{
table_3_8_parameter_estimates(base_dir = "mcmc_outputs/zimphia")
}
\arguments{
\item{base_dir}{ZIMPHIA results directory. Default "mcmc_outputs/zimphia".}
}
\value{
Character representation of the gt table in LaTeX format.
}
\description{
Displays posterior medians and 95\% CrIs for core parameters with MH and
HMC side-by-side.
}
