% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_helpers.R, R/viz_survival.R
\name{validate_sim_data}
\alias{validate_sim_data}
\title{Validate simulation data}
\usage{
validate_sim_data(df, require_cols = c("L", "R", "weight"))

validate_sim_data(df, require_cols = c("L", "R", "weight"))
}
\arguments{
\item{df}{Data.frame to validate}

\item{require_cols}{Character vector of required column names}
}
\value{
Logical, TRUE if valid

Logical, TRUE if valid
}
\description{
Validate simulation data

Validate simulation data
}
\keyword{internal}
