% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{load_diagnostics}
\alias{load_diagnostics}
\title{Load all diagnostic files for a given method and sample size}
\usage{
load_diagnostics(method, n_obs, results_dir, verbose = TRUE)
}
\arguments{
\item{method}{Character, either "hmc" or "mh"}

\item{n_obs}{Integer, sample size (200, 2000, or 10000)}

\item{results_dir}{Path to results directory}

\item{verbose}{Print progress messages?}
}
\value{
Tibble with all diagnostics
}
\description{
Load all diagnostic files for a given method and sample size
}
\keyword{internal}
